Relocate AuctionService project files to 'src' directory
This commit moves files associated with the AuctionService project to a new 'src' directory. The intention behind this change is for better organization of project files and a clean separation of the application's source code from other elements within the project structure. Adjustments have been made to paths inside file contents and the solution file to reflect the new file locations.