Integrate MassTransit and add new consumers
MassTransit has been integrated into the AuctionService along with the addition of consumers for various auction events. The AuctionService now has the capability to publish messages and participate in distributed transactions via MassTransit. Simultaneously, the SearchService is configured to consume the auction messages and respond to events such as an auction being updated, deleted or finished. Code has been added to handle these events and modify data accordingly. Additionally, MassTransit settings were adjusted with PostgreSQL as the database for the Outbox, automatic endpoint configuration, and kebab-case for endpoint names.