Add new configurations for Let's Encrypt and update auth options
Two new configuration files for the Let's Encrypt certificates were added for both staging and production environments. Their purpose is to ensure secure connections. Additionally, the authentication options file was updated to enforce the presence of an Identity Server URL in the environment variables. The setup for the ingress service in the Kubernetes infrastructure was also updated to include the annotations necessary for the Cert-Manager to function properly